refactor(engine): Break main router into smaller routers
