+++ tracecat/api/app.py
@@ -1,86 +1,29 @@
-import asyncio
 from contextlib import asynccontextmanager
-from typing import Annotated, Any
 
-import orjson
-from fastapi import (
-    Depends,
-    FastAPI,
-    File,
-    Header,
-    HTTPException,
-    Query,
-    Request,
-    UploadFile,
-    status,
-)
+from fastapi import FastAPI, Request, status
 from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import ORJSONResponse
 from fastapi.routing import APIRoute
-from pydantic import ValidationError
-from sqlalchemy import or_
-from sqlalchemy.exc import NoResultFound
-from sqlmodel import Session, select
 
-from tracecat import config, identifiers, validation
-from tracecat.auth.credentials import (
-    TemporaryRole,
-    authenticate_service,
-    authenticate_user,
-    authenticate_user_or_service,
-)
+from tracecat import config
+from tracecat.api.routers.actions import router as actions_router
+from tracecat.api.routers.cases.actions import router as case_actions_router
+from tracecat.api.routers.cases.contexts import router as case_contexts_router
+from tracecat.api.routers.cases.management import router as cases_router
+from tracecat.api.routers.public.callbacks import router as callback_router
+from tracecat.api.routers.public.webhooks import router as webhook_router
+from tracecat.api.routers.schedules import router as schedules_router
+from tracecat.api.routers.secrets import router as secrets_router
+from tracecat.api.routers.udfs import router as udfs_router
+from tracecat.api.routers.users import router as users_router
+from tracecat.api.routers.validation import router as validation_router
 from tracecat.contexts import ctx_role
-from tracecat.db.engine import get_session, get_session_context_manager, initialize_db
-from tracecat.db.schemas import (
-    Action,
-    Case,
-    CaseAction,
-    CaseContext,
-    CaseEvent,
-    Schedule,
-    Secret,
-    UDFSpec,
-    User,
-    Webhook,
-    WorkflowDefinition,
-)
-from tracecat.dsl import schedules
-from tracecat.dsl.common import DSLInput
-
-# TODO: Clean up API params / response "zoo"
-# lots of repetition and inconsistency
-from tracecat.dsl.validation import validate_trigger_inputs
+from tracecat.db.engine import initialize_db
 from tracecat.logging import logger
 from tracecat.middleware import RequestLoggingMiddleware
-from tracecat.parse import parse_child_webhook
-from tracecat.secrets.service import SecretsService
-from tracecat.types.api import (
-    ActionMetadataResponse,
-    ActionResponse,
-    CaseActionParams,
-    CaseContextParams,
-    CaseEventParams,
-    CaseParams,
-    CaseResponse,
-    CreateActionParams,
-    CreateScheduleParams,
-    CreateSecretParams,
-    SearchScheduleParams,
-    SearchSecretsParams,
-    SecretResponse,
-    ServiceCallbackAction,
-    UDFArgsValidationResponse,
-    UpdateActionParams,
-    UpdateScheduleParams,
-    UpdateSecretParams,
-    UpdateUserParams,
-)
-from tracecat.types.auth import Role
-from tracecat.types.exceptions import TracecatException, TracecatValidationError
-from tracecat.workflow.executions.models import CreateWorkflowExecutionResponse
+from tracecat.types.exceptions import TracecatException
 from tracecat.workflow.executions.router import router as workflow_executions_router
-from tracecat.workflow.executions.service import WorkflowExecutionsService
 from tracecat.workflow.management.router import router as workflow_management_router
 
 
@@ -175,9 +118,22 @@ def create_app(**kwargs) -> FastAPI:
         **kwargs,
     )
     app.logger = logger
+
     # Routers
+    app.include_router(webhook_router)
+    app.include_router(callback_router)
     app.include_router(workflow_management_router)
     app.include_router(workflow_executions_router)
+    app.include_router(actions_router)
+    app.include_router(udfs_router)
+    app.include_router(cases_router)
+    app.include_router(case_actions_router)
+    app.include_router(case_contexts_router)
+    app.include_router(secrets_router)
+    app.include_router(schedules_router)
+    app.include_router(users_router)
+    app.include_router(validation_router)
+
     # Exception handlers
     app.add_exception_handler(Exception, generic_exception_handler)
     app.add_exception_handler(TracecatException, tracecat_exception_handler)
@@ -200,9 +156,6 @@ def create_app(**kwargs) -> FastAPI:
 app = create_app()
 
 
-# ----- Utility ----- #
-
-
 @app.get("/", include_in_schema=False)
 def root() -> dict[str, str]:
     return {"message": "Hello world. I am the API."}
@@ -211,1433 +164,3 @@ def root() -> dict[str, str]:
 @app.get("/health", tags=["public"])
 def check_health() -> dict[str, str]:
     return {"message": "Hello world. I am the API. This is the health endpoint."}
-
-
-# ----- Trigger handlers ----- #
-
-
-def validate_incoming_webhook(
-    webhook_id: str,
-    secret: str,
-    request: Request,
-    *,
-    validate_method: bool = True,
-) -> WorkflowDefinition:
-    """Validate incoming webhook request.
-
-    NOte: The webhook ID here is the workflow ID.
-    """
-    with get_session_context_manager() as session:
-        result = session.exec(select(Webhook).where(Webhook.workflow_id == webhook_id))
-        try:
-            # One webhook per workflow
-            webhook = result.one()
-        except NoResultFound as e:
-            logger.opt(exception=e).error("Webhook does not exist", error=e)
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="Unauthorized webhook request.",
-            ) from e
-
-        if secret != webhook.secret:
-            logger.error("Secret does not match")
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="Unauthorized webhook request",
-            )
-
-        # If we're here, the webhook has been validated
-        if webhook.status == "offline":
-            logger.error("Webhook is offline")
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Webhook is offline",
-            )
-
-        if validate_method and webhook.method.lower() != request.method.lower():
-            logger.error("Method does not match")
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Request method not allowed",
-            )
-
-        # Reaching here means the webhook is online and connected to an entrypoint
-
-        # Match the webhook id with the workflow id and get the latest version
-        # of the workflow defitniion.
-        result = session.exec(
-            select(WorkflowDefinition)
-            .where(WorkflowDefinition.workflow_id == webhook_id)
-            .order_by(WorkflowDefinition.version.desc())
-        )
-        try:
-            defn = result.first()
-            if not defn:
-                raise NoResultFound(
-                    "No workflow definition found for workflow ID. Please commit your changes to the workflow and try again."
-                )
-        except NoResultFound as e:
-            # No workflow associated with the webhook
-            logger.error(str(e), error=e)
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detail=str(e)
-            ) from e
-
-        # Check if the workflow is active
-
-        if defn.workflow.status == "offline":
-            logger.error("Workflow is offline")
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Workflow is offline",
-            )
-
-        # If we are here, all checks have passed
-        return WorkflowDefinition.model_validate(defn)
-
-
-async def handle_incoming_webhook(
-    request: Request, path: str, secret: str, validate_method: bool = True
-) -> WorkflowDefinition:
-    """Handle an incoming webhook request and set the Role context."""
-    # TODO(perf): Replace this when we get async sessions
-    with logger.contextualize(webhook_id=path):
-        defn = await asyncio.to_thread(
-            validate_incoming_webhook,
-            webhook_id=path,
-            secret=secret,
-            request=request,
-            validate_method=validate_method,
-        )
-    ctx_role.set(
-        Role(type="service", user_id=defn.owner_id, service_id="tracecat-runner")
-    )
-    return defn
-
-
-@app.post("/webhooks/{path}/{secret}", tags=["public"])
-async def incoming_webhook(
-    defn: Annotated[WorkflowDefinition, Depends(handle_incoming_webhook)],
-    path: str,
-    payload: dict[str, Any] | None = None,
-    x_tracecat_enable_runtime_tests: Annotated[str | None, Header()] = None,
-) -> CreateWorkflowExecutionResponse:
-    """
-    Webhook endpoint to trigger a workflow.
-
-    This is an external facing endpoint is used to trigger a workflow by sending a webhook request.
-    The workflow is identified by the `path` parameter, which is equivalent to the workflow id.
-    """
-    logger.info(
-        "Webhook hit",
-        path=path,
-        payload=payload,
-        role=ctx_role.get(),
-    )
-
-    dsl_input = DSLInput(**defn.content)
-
-    enable_runtime_tests = (x_tracecat_enable_runtime_tests or "false").lower() in (
-        "1",
-        "true",
-    )
-
-    service = await WorkflowExecutionsService.connect()
-    response = service.create_workflow_execution_nowait(
-        dsl=dsl_input,
-        wf_id=path,
-        payload=payload,
-        enable_runtime_tests=enable_runtime_tests,
-    )
-    return response
-
-
-async def handle_service_callback(
-    request: Request, service: str
-) -> ServiceCallbackAction | None:
-    if service == "slack":
-        # if (
-        #     request.headers["user-agent"]
-        #     != "Slackbot 1.0 (+https://api.slack.com/robots)"
-        # ):
-        #     raise HTTPException(
-        #         status_code=status.HTTP_400_BAD_REQUEST,
-        #         detail="Invalid User-Agent",
-        #     )
-        # NOTE: This coroutine can only be consumed once!
-        form_data = await request.form()
-        json_payload = form_data.get("payload")
-        payload = orjson.loads(json_payload)
-        # Extract out the webhook
-        if (actions := payload.get("actions")) is None:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Slack callback: Invalid payload",
-            )
-        logger.info("Received slack action", actions=actions)
-        if len(actions) < 1:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Slack callback: No actions",
-            )
-        action = actions[0]
-        match action:
-            case {
-                "type": "static_select",
-                "action_id": url,
-                "selected_option": {"value": kv_params},
-            }:
-                # e.g.
-                # {
-                #     "type": "static_select",
-                #     "action_id": "...",
-                #     "block_id": "nMMIK",
-                #     "selected_option": {
-                #         "text": {
-                #             "type": "plain_text",
-                #             "text": "True Positive",
-                #             "emoji": True,
-                #         },
-                #         "value": '["closed", "true_positive"]',
-                #     },
-                #     "action_ts": "1719281272.742854",
-                # }
-
-                logger.info("Matched static select action", action=action)
-                child_wh = parse_child_webhook(url, [kv_params])
-
-                if not child_wh:
-                    raise HTTPException(
-                        status_code=status.HTTP_400_BAD_REQUEST,
-                        detail="Invalid child webhook URL",
-                    )
-                return ServiceCallbackAction(
-                    action="webhook",
-                    payload=child_wh["payload"],
-                    metadata={"path": child_wh["path"], "secret": child_wh["secret"]},
-                )
-
-            case {"type": action_type}:
-                logger.error(
-                    "Invalid action type", action_type=action_type, action=action
-                )
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail=f"Invalid action type {action_type}",
-                )
-            case _:
-                logger.error("Invalid action", action=action)
-                raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail="Invalid action",
-                )
-    return None
-
-
-@app.post("/callback/{service}", tags=["public"])
-async def webhook_callback(
-    request: Request,
-    service: str,
-    next_action: Annotated[
-        ServiceCallbackAction | None, Depends(handle_service_callback)
-    ],
-) -> dict[str, str]:
-    """Receive a callback from an external service.
-
-    This can be used to trigger a workflow from an external service, or perform some other actions.
-    """
-
-    match next_action:
-        case ServiceCallbackAction(
-            action="webhook",
-            payload=payload,
-            metadata={"path": path, "secret": secret},
-        ):
-            # Don't validate method because callback is always POST
-            defn = await handle_incoming_webhook(
-                request, path, secret, validate_method=False
-            )
-            logger.info(
-                "Received Webhook in callback",
-                service=service,
-                path=path,
-                payload=payload,
-                role=ctx_role.get(),
-            )
-
-            # Fetch the DSL from the workflow object
-            dsl_input = DSLInput(**defn.content)
-
-            wf_exec_service = await WorkflowExecutionsService.connect()
-            response = wf_exec_service.create_workflow_execution_nowait(
-                dsl=dsl_input,
-                wf_id=path,
-                payload=payload,
-            )
-            return {
-                "status": "ok",
-                "message": "Webhook callback processed",
-                "service": service,
-                "details": response,
-            }
-
-        case None:
-            logger.info("No next action", service=service)
-            return {"status": "ok", "message": "No action taken", "service": service}
-        case _:
-            logger.error("Unsupported next action", next_action=next_action)
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Unsupported next action in webhook callback for {service!r} service",
-            )
-
-
-# ----- Workflow Schedules ----- #
-
-
-@app.get("/schedules", tags=["schedules"])
-async def list_schedules(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: identifiers.WorkflowID | None = None,
-    session: Session = Depends(get_session),
-) -> list[Schedule]:
-    """List all schedules for a workflow."""
-    statement = select(Schedule).where(Schedule.owner_id == role.user_id)
-    if workflow_id:
-        statement = statement.where(Schedule.workflow_id == workflow_id)
-    result = session.exec(statement)
-    try:
-        return result.all()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-
-
-@app.post("/schedules", tags=["schedules"])
-async def create_schedule(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: CreateScheduleParams,
-    session: Session = Depends(get_session),
-) -> Schedule:
-    """Create a schedule for a workflow."""
-
-    with logger.contextualize(role=role):
-        result = session.exec(
-            select(WorkflowDefinition)
-            .where(WorkflowDefinition.workflow_id == params.workflow_id)
-            .order_by(WorkflowDefinition.version.desc())
-        )
-        try:
-            if not (defn_data := result.first()):
-                raise NoResultFound("No workflow definition found for workflow ID")
-        except NoResultFound as e:
-            logger.opt(exception=e).error("Invalid workflow ID", error=e)
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detail="Invalid workflow ID"
-            ) from e
-
-        schedule = Schedule(
-            owner_id=role.user_id, **params.model_dump(exclude_unset=True)
-        )
-        session.refresh(defn_data)
-        defn = WorkflowDefinition.model_validate(defn_data)
-        dsl = DSLInput(**defn.content)
-        if params.inputs:
-            dsl.trigger_inputs = params.inputs
-
-        try:
-            # Set the role for the schedule as the tracecat-runner
-            with TemporaryRole(
-                type="service",
-                user_id=defn.owner_id,
-                service_id="tracecat-schedule-runner",
-            ) as sch_role:
-                handle = await schedules.create_schedule(
-                    workflow_id=params.workflow_id,
-                    schedule_id=schedule.id,
-                    dsl=dsl,
-                    every=params.every,
-                    offset=params.offset,
-                    start_at=params.start_at,
-                    end_at=params.end_at,
-                )
-                logger.info(
-                    "Created schedule",
-                    handle_id=handle.id,
-                    workflow_id=params.workflow_id,
-                    schedule_id=schedule.id,
-                    sch_role=sch_role,
-                )
-
-            session.add(schedule)
-            session.commit()
-            session.refresh(schedule)
-            return schedule
-        except Exception as e:
-            session.rollback()
-            logger.opt(exception=e).error("Error creating schedule", error=e)
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail="Error creating schedule",
-            ) from e
-
-
-@app.get("/schedules/{schedule_id}", tags=["schedules"])
-def get_schedule(
-    role: Annotated[Role, Depends(authenticate_user)],
-    schedule_id: identifiers.ScheduleID,
-    session: Session = Depends(get_session),
-) -> Schedule:
-    """Get a schedule from a workflow."""
-    statement = select(Schedule).where(
-        Schedule.owner_id == role.user_id, Schedule.id == schedule_id
-    )
-    result = session.exec(statement)
-    try:
-        return result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-
-
-@app.post("/schedules/{schedule_id}", tags=["schedules"])
-async def update_schedule(
-    role: Annotated[Role, Depends(authenticate_user)],
-    schedule_id: identifiers.ScheduleID,
-    params: UpdateScheduleParams,
-    session: Session = Depends(get_session),
-) -> Schedule:
-    """Update a schedule from a workflow. You cannot update the Workflow Definition, but you can update other fields."""
-    statement = select(Schedule).where(
-        Schedule.owner_id == role.user_id, Schedule.id == schedule_id
-    )
-    result = session.exec(statement)
-    try:
-        schedule = result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-
-    try:
-        # (1) Synchronize with Temporal
-        await schedules.update_schedule(schedule_id, params)
-
-        # (2) Update the schedule
-        for key, value in params.model_dump(exclude_unset=True).items():
-            # Safety: params have been validated
-            setattr(schedule, key, value)
-
-        session.add(schedule)
-        session.commit()
-        session.refresh(schedule)
-        return schedule
-    except Exception as e:
-        session.rollback()
-        logger.opt(exception=e).error("Error creating schedule", error=e)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Error creating schedule",
-        ) from e
-
-
-@app.delete(
-    "/schedules/{schedule_id}",
-    status_code=status.HTTP_204_NO_CONTENT,
-    tags=["schedules"],
-)
-async def delete_schedule(
-    role: Annotated[Role, Depends(authenticate_user)],
-    schedule_id: identifiers.ScheduleID,
-    session: Session = Depends(get_session),
-) -> None:
-    """Delete a schedule from a workflow."""
-    statement = select(Schedule).where(
-        Schedule.owner_id == role.user_id, Schedule.id == schedule_id
-    )
-    result = session.exec(statement)
-    schedule = result.one_or_none()
-    if not schedule:
-        logger.warning(
-            "Schedule not found, attempt to delete underlying Temporal schedule...",
-            schedule_id=schedule_id,
-        )
-
-    try:
-        # Delete the schedule from Temporal first
-        await schedules.delete_schedule(schedule_id)
-
-        # If successful, delete the schedule from the database
-        if schedule:
-            session.delete(schedule)
-            session.commit()
-        else:
-            logger.warning(
-                "Schedule was already deleted from the database",
-                schedule_id=schedule_id,
-            )
-    except Exception as e:
-        logger.error("Error deleting schedule", error=e)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Error deleting schedule",
-        ) from e
-
-
-@app.get("/schedules/search", tags=["schedules"])
-def search_schedules(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: SearchScheduleParams,
-    session: Session = Depends(get_session),
-) -> list[Schedule]:
-    """**[WORK IN PROGRESS]** Search for schedules."""
-    statement = select(Schedule).where(Schedule.owner_id == role.user_id)
-    results = session.exec(statement)
-    schedules = results.all()
-    return schedules
-
-
-# ----- Actions ----- #
-
-
-@app.get("/actions", tags=["actions"])
-def list_actions(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    session: Session = Depends(get_session),
-) -> list[ActionMetadataResponse]:
-    """List all actions for a workflow."""
-    statement = select(Action).where(
-        Action.owner_id == role.user_id,
-        Action.workflow_id == workflow_id,
-    )
-    results = session.exec(statement)
-    actions = results.all()
-    action_metadata = [
-        ActionMetadataResponse(
-            id=action.id,
-            workflow_id=workflow_id,
-            type=action.type,
-            title=action.title,
-            description=action.description,
-            status=action.status,
-            key=action.key,
-        )
-        for action in actions
-    ]
-    return action_metadata
-
-
-@app.post("/actions", tags=["actions"])
-def create_action(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: CreateActionParams,
-    session: Session = Depends(get_session),
-) -> ActionMetadataResponse:
-    """Create a new action for a workflow."""
-    action = Action(
-        owner_id=role.user_id,
-        workflow_id=params.workflow_id,
-        type=params.type,
-        title=params.title,
-        description="",  # Default to empty string
-    )
-    # Check if a clashing action ref exists
-    statement = select(Action).where(
-        Action.owner_id == role.user_id,
-        Action.workflow_id == action.workflow_id,
-        Action.ref == action.ref,
-    )
-    if session.exec(statement).first():
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="Action ref already exists in the workflow",
-        )
-
-    session.add(action)
-    session.commit()
-    session.refresh(action)
-
-    action_metadata = ActionMetadataResponse(
-        id=action.id,
-        workflow_id=params.workflow_id,
-        type=params.type,
-        title=action.title,
-        description=action.description,
-        status=action.status,
-        key=action.key,
-    )
-    return action_metadata
-
-
-@app.get("/actions/{action_id}", tags=["actions"])
-def get_action(
-    role: Annotated[Role, Depends(authenticate_user)],
-    action_id: str,
-    workflow_id: str,
-    session: Session = Depends(get_session),
-) -> ActionResponse:
-    """Get an action."""
-    statement = select(Action).where(
-        Action.owner_id == role.user_id,
-        Action.id == action_id,
-        Action.workflow_id == workflow_id,
-    )
-    result = session.exec(statement)
-    try:
-        action = result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-
-    return ActionResponse(
-        id=action.id,
-        type=action.type,
-        title=action.title,
-        description=action.description,
-        status=action.status,
-        inputs=action.inputs,
-        key=action.key,
-        control_flow=action.control_flow,
-    )
-
-
-@app.post("/actions/{action_id}", tags=["actions"])
-def update_action(
-    role: Annotated[Role, Depends(authenticate_user)],
-    action_id: str,
-    params: UpdateActionParams,
-    session: Session = Depends(get_session),
-) -> ActionResponse:
-    """Update an action."""
-    # Fetch the action by id
-    statement = select(Action).where(
-        Action.owner_id == role.user_id,
-        Action.id == action_id,
-    )
-    result = session.exec(statement)
-    try:
-        action = result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-
-    if params.title is not None:
-        action.title = params.title
-    if params.description is not None:
-        action.description = params.description
-    if params.status is not None:
-        action.status = params.status
-    if params.inputs is not None:
-        action.inputs = params.inputs
-    if params.control_flow is not None:
-        action.control_flow = params.control_flow.model_dump(mode="json")
-
-    session.add(action)
-    session.commit()
-    session.refresh(action)
-
-    return ActionResponse(
-        id=action.id,
-        type=action.type,
-        title=action.title,
-        description=action.description,
-        status=action.status,
-        inputs=action.inputs,
-        key=action.key,
-    )
-
-
-@app.delete(
-    "/actions/{action_id}", status_code=status.HTTP_204_NO_CONTENT, tags=["actions"]
-)
-def delete_action(
-    role: Annotated[Role, Depends(authenticate_user)],
-    action_id: str,
-    session: Session = Depends(get_session),
-) -> None:
-    """Delete an action."""
-    statement = select(Action).where(
-        Action.owner_id == role.user_id,
-        Action.id == action_id,
-    )
-    result = session.exec(statement)
-    try:
-        action = result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-    # If the user doesn't own this workflow, they can't delete the action
-    session.delete(action)
-    session.commit()
-
-
-# ----- Case Management ----- #
-
-
-@app.post(
-    "/workflows/{workflow_id}/cases",
-    status_code=status.HTTP_201_CREATED,
-    tags=["cases"],
-)
-def create_case(
-    role: Annotated[Role, Depends(authenticate_service)],  # M2M
-    workflow_id: str,
-    cases: list[CaseParams],
-    session: Session = Depends(get_session),
-) -> CaseResponse:
-    """Create a new case for a workflow."""
-    case = Case(
-        owner_id=role.user_id,
-        workflow_id=workflow_id,
-        **cases.model_dump(),
-    )
-    session.add(case)
-    session.commit()
-    session.refresh(case)
-
-    return CaseResponse(
-        id=case.id,
-        owner_id=case.owner_id,
-        created_at=case.created_at,
-        updated_at=case.updated_at,
-        workflow_id=case.workflow_id,
-        case_title=case.case_title,
-        payload=case.payload,
-        malice=case.malice,
-        status=case.status,
-        priority=case.priority,
-        action=case.action,
-        context=case.context,
-        tags=case.tags,
-    )
-
-
-@app.get("/workflows/{workflow_id}/cases", tags=["cases"])
-def list_cases(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    limit: int = 100,
-    session: Session = Depends(get_session),
-) -> list[CaseResponse]:
-    """List all cases for a workflow."""
-    query = select(Case).where(
-        Case.owner_id == role.user_id, Case.workflow_id == workflow_id
-    )
-    result = session.exec(query)
-    try:
-        cases = result.fetchmany(size=limit)
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-
-    return [
-        CaseResponse(
-            id=case.id,
-            owner_id=case.owner_id,
-            created_at=case.created_at,
-            updated_at=case.updated_at,
-            workflow_id=case.workflow_id,
-            case_title=case.case_title,
-            payload=case.payload,
-            malice=case.malice,
-            status=case.status,
-            priority=case.priority,
-            action=case.action,
-            context=case.context,
-            tags=case.tags,
-        )
-        for case in cases
-    ]
-
-
-@app.get("/workflows/{workflow_id}/cases/{case_id}", tags=["cases"])
-def get_case(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    case_id: str,
-    session: Session = Depends(get_session),
-) -> CaseResponse:
-    """Get a specific case for a workflow."""
-    query = select(Case).where(
-        Case.owner_id == role.user_id,
-        Case.workflow_id == workflow_id,
-        Case.id == case_id,
-    )
-    case = session.exec(query).one_or_none()
-    if case is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        )
-    return CaseResponse(
-        id=case.id,
-        owner_id=case.owner_id,
-        created_at=case.created_at,
-        updated_at=case.updated_at,
-        workflow_id=case.workflow_id,
-        case_title=case.case_title,
-        payload=case.payload,
-        malice=case.malice,
-        status=case.status,
-        priority=case.priority,
-        action=case.action,
-        context=case.context,
-        tags=case.tags,
-    )
-
-
-@app.post("/workflows/{workflow_id}/cases/{case_id}", tags=["cases"])
-def update_case(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    case_id: str,
-    params: CaseParams,
-    session: Session = Depends(get_session),
-) -> CaseResponse:
-    """Update a specific case for a workflow."""
-    query = select(Case).where(
-        Case.owner_id == role.user_id,
-        Case.workflow_id == workflow_id,
-        Case.id == case_id,
-    )
-    case = session.exec(query).one_or_none()
-    if case is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        )
-
-    for key, value in params.model_dump(exclude_unset=True).items():
-        # Safety: params have been validated
-        setattr(case, key, value)
-
-    session.add(case)
-    session.commit()
-    session.refresh(case)
-    return CaseResponse(
-        id=case.id,
-        owner_id=case.owner_id,
-        created_at=case.created_at,
-        updated_at=case.updated_at,
-        workflow_id=case.workflow_id,
-        case_title=case.case_title,
-        payload=case.payload,
-        malice=case.malice,
-        status=case.status,
-        priority=case.priority,
-        action=case.action,
-        context=case.context,
-        tags=case.tags,
-    )
-
-
-@app.post(
-    "/workflows/{workflow_id}/cases/{case_id}/events",
-    status_code=status.HTTP_201_CREATED,
-    tags=["cases"],
-)
-def create_case_event(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    case_id: str,
-    params: CaseEventParams,
-    session: Session = Depends(get_session),
-) -> None:
-    """Create a new Case Event."""
-    case_event = CaseEvent(
-        owner_id=role.user_id,
-        case_id=case_id,
-        workflow_id=workflow_id,
-        initiator_role=role.type,
-        **params.model_dump(),
-    )
-    session.add(case_event)
-    session.commit()
-    session.refresh(case_event)
-    return case_event
-
-
-@app.get("/workflows/{workflow_id}/cases/{case_id}/events", tags=["cases"])
-def list_case_events(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    case_id: str,
-    session: Session = Depends(get_session),
-) -> list[CaseEvent]:
-    """List all Case Events."""
-    query = select(CaseEvent).where(
-        CaseEvent.owner_id == role.user_id,
-        CaseEvent.workflow_id == workflow_id,
-        CaseEvent.case_id == case_id,
-    )
-    case_events = session.exec(query).all()
-    return case_events
-
-
-@app.get("/workflows/{workflow_id}/cases/{case_id}/events/{event_id}", tags=["cases"])
-def get_case_event(
-    role: Annotated[Role, Depends(authenticate_user)],
-    workflow_id: str,
-    case_id: str,
-    event_id: str,
-    session: Session = Depends(get_session),
-):
-    """Get a specific case event."""
-    query = select(CaseEvent).where(
-        CaseEvent.owner_id == role.user_id,
-        CaseEvent.workflow_id == workflow_id,
-        CaseEvent.case_id == case_id,
-        CaseEvent.id == event_id,
-    )
-    case_event = session.exec(query).one_or_none()
-    if case_event is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        )
-    return case_event
-
-
-# ----- Available Case Actions ----- #
-
-
-@app.get("/case-actions", tags=["cases"])
-def list_case_actions(
-    role: Annotated[Role, Depends(authenticate_user)],
-    session: Session = Depends(get_session),
-) -> list[CaseAction]:
-    """List all case actions."""
-    statement = select(CaseAction).where(
-        or_(
-            CaseAction.owner_id == "tracecat",
-            CaseAction.owner_id == role.user_id,
-        )
-    )
-    actions = session.exec(statement).all()
-    return actions
-
-
-@app.post("/case-actions", tags=["cases"])
-def create_case_action(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: CaseActionParams,
-    session: Session = Depends(get_session),
-) -> CaseAction:
-    """Create a new case action."""
-    case_action = CaseAction(owner_id=role.user_id, **params.model_dump())
-    session.add(case_action)
-    session.commit()
-    session.refresh(case_action)
-    return case_action
-
-
-@app.delete("/case-actions/{case_action_id}", tags=["cases"])
-def delete_case_action(
-    role: Annotated[Role, Depends(authenticate_user)],
-    case_action_id: str,
-    session: Session = Depends(get_session),
-):
-    """Delete a case action."""
-    statement = select(CaseAction).where(
-        CaseAction.owner_id == role.user_id,
-        CaseAction.id == case_action_id,
-    )
-    result = session.exec(statement)
-    try:
-        action = result.one()
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-    session.delete(action)
-    session.commit()
-
-
-# ----- Available Context Labels ----- #
-
-
-@app.get("/case-contexts", tags=["cases"])
-def list_case_contexts(
-    role: Annotated[Role, Depends(authenticate_user)],
-    session: Session = Depends(get_session),
-) -> list[CaseContext]:
-    """List all case contexts."""
-    statement = select(CaseContext).where(
-        or_(
-            CaseContext.owner_id == "tracecat",
-            CaseContext.owner_id == role.user_id,
-        )
-    )
-    actions = session.exec(statement).all()
-    return actions
-
-
-@app.post("/case-contexts", tags=["cases"])
-def create_case_context(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: CaseContextParams,
-    session: Session = Depends(get_session),
-) -> CaseContext:
-    """Create a new case context."""
-    case_context = CaseContext(owner_id=role.user_id, **params.model_dump())
-    session.add(case_context)
-    session.commit()
-    session.refresh(case_context)
-    return params
-
-
-@app.delete("/case-contexts/{case_context_id}", tags=["cases"])
-def delete_case_context(
-    role: Annotated[Role, Depends(authenticate_user)],
-    case_context_id: str,
-    session: Session = Depends(get_session),
-):
-    """Delete a case context."""
-    statement = select(CaseContext).where(
-        CaseContext.owner_id == role.user_id,
-        CaseContext.id == case_context_id,
-    )
-    result = session.exec(statement)
-    try:
-        action = result.one()
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
-        ) from e
-    session.delete(action)
-    session.delete(action)
-    session.commit()
-
-
-# ----- Users ----- #
-
-
-@app.post("/users", status_code=status.HTTP_201_CREATED, tags=["users"])
-def create_user(
-    role: Annotated[Role, Depends(authenticate_user)],
-    session: Session = Depends(get_session),
-) -> User:
-    """Create new user."""
-
-    # Check if user exists
-    statement = select(User).where(User.id == role.user_id).limit(1)
-    result = session.exec(statement)
-
-    user = result.one_or_none()
-    if user is not None:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT, detail="User already exists"
-        )
-    user = User(owner_id="tracecat", id=role.user_id)
-
-    session.add(user)
-    session.commit()
-    session.refresh(user)
-    return user
-
-
-@app.get("/users", tags=["users"])
-def get_user(
-    role: Annotated[Role, Depends(authenticate_user)],
-    session: Session = Depends(get_session),
-) -> User:
-    """Get a user."""
-
-    # Get user given user_id
-    statement = select(User).where(User.id == role.user_id)
-    result = session.exec(statement)
-    try:
-        user = result.one()
-        return user
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
-        ) from e
-
-
-@app.post("/users", status_code=status.HTTP_204_NO_CONTENT, tags=["users"])
-def update_user(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: UpdateUserParams,
-    session: Session = Depends(get_session),
-) -> None:
-    """Update a user."""
-
-    statement = select(User).where(User.id == role.user_id)
-    result = session.exec(statement)
-    try:
-        user = result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
-        ) from e
-
-    if params.tier is not None:
-        user.tier = params.tier
-    if params.settings is not None:
-        user.settings = params.settings
-
-    session.add(user)
-    session.commit()
-
-
-@app.delete("/users", status_code=status.HTTP_204_NO_CONTENT, tags=["users"])
-def delete_user(
-    role: Annotated[Role, Depends(authenticate_user)],
-    session: Session = Depends(get_session),
-) -> None:
-    """Delete a user."""
-
-    statement = select(User).where(User.id == role.user_id)
-    result = session.exec(statement)
-    try:
-        user = result.one()
-    except NoResultFound as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
-        ) from e
-    session.delete(user)
-    session.commit()
-
-
-# ----- Secrets ----- #
-
-
-@app.get("/secrets", tags=["secrets"])
-def list_secrets(
-    role: Annotated[Role, Depends(authenticate_user)],
-    session: Session = Depends(get_session),
-) -> list[SecretResponse]:
-    """List user secrets."""
-    service = SecretsService(session, role)
-    secrets = service.list_secrets()
-    return [
-        SecretResponse(
-            id=secret.id,
-            type=secret.type,
-            name=secret.name,
-            description=secret.description,
-            keys=[kv.key for kv in service.decrypt_keys(secret.encrypted_keys)],
-        )
-        for secret in secrets
-    ]
-
-
-@app.get("/secrets/{secret_name}", tags=["secrets"])
-def get_secret(
-    # NOTE(auth): Worker service can also access secrets
-    role: Annotated[Role, Depends(authenticate_user_or_service)],
-    secret_name: str,
-    session: Session = Depends(get_session),
-) -> Secret:
-    """Get a secret."""
-
-    # Check if secret exists
-    statement = (
-        select(Secret)
-        .where(Secret.owner_id == role.user_id, Secret.name == secret_name)
-        .limit(1)
-    )
-    result = session.exec(statement)
-    secret = result.one_or_none()
-    if secret is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Secret not found"
-        )
-    # NOTE: IMPLICIT TYPE COERCION
-    # Encrypted keys as bytes gets cast a string as to be JSON serializable
-    return secret
-
-
-@app.post("/secrets", status_code=status.HTTP_201_CREATED, tags=["secrets"])
-def create_secret(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: CreateSecretParams,
-    session: Session = Depends(get_session),
-) -> None:
-    """Create a secret."""
-    service = SecretsService(session, role)
-    secret = service.get_secret_by_name(params.name)
-    if secret is not None:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT, detail="Secret already exists"
-        )
-
-    new_secret = Secret(
-        owner_id=role.user_id,
-        name=params.name,
-        type=params.type,
-        description=params.description,
-        tags=params.tags,
-        encrypted_keys=service.encrypt_keys(params.keys),
-    )
-    service.create_secret(new_secret)
-
-
-@app.post(
-    "/secrets/{secret_name}", status_code=status.HTTP_201_CREATED, tags=["secrets"]
-)
-def update_secret(
-    role: Annotated[Role, Depends(authenticate_user)],
-    secret_name: str,
-    params: UpdateSecretParams,
-    session: Session = Depends(get_session),
-) -> Secret:
-    """Update a secret"""
-    service = SecretsService(session, role)
-    secret = service.get_secret_by_name(secret_name)
-    if secret is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Secret does not exist"
-        )
-    maybe_clashing_secret = service.get_secret_by_name(params.name)
-    if maybe_clashing_secret is not None and maybe_clashing_secret.id != secret.id:
-        name = maybe_clashing_secret.name
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=f"Secret with name {name} already exists",
-        )
-    return service.update_secret(secret, params)
-
-
-@app.delete(
-    "/secrets/{secret_name}", status_code=status.HTTP_204_NO_CONTENT, tags=["secrets"]
-)
-def delete_secret(
-    role: Annotated[Role, Depends(authenticate_user)],
-    secret_name: str,
-    session: Session = Depends(get_session),
-) -> None:
-    """Delete a secret."""
-    service = SecretsService(session, role)
-    secret = service.get_secret_by_name(secret_name)
-    if secret is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Secret does not exist"
-        )
-    service.delete_secret(secret)
-
-
-@app.post("/secrets/search", tags=["secrets"])
-def search_secrets(
-    role: Annotated[Role, Depends(authenticate_user)],
-    params: SearchSecretsParams,
-    session: Session = Depends(get_session),
-) -> list[Secret]:
-    """**[WORK IN PROGRESS]**   Get a secret by ID."""
-    statement = (
-        select(Secret)
-        .where(Secret.owner_id == role.user_id)
-        .filter(*[Secret.name == name for name in params.names])
-    )
-    result = session.exec(statement)
-    secrets = result.all()
-    return secrets
-
-
-# ----- UDFs ----- #
-
-
-@app.get("/udfs", tags=["udfs"])
-def list_udfs(
-    role: Annotated[Role, Depends(authenticate_user_or_service)],
-    limit: int | None = None,
-    ns: list[str] | None = Query(None),
-    session: Session = Depends(get_session),
-) -> list[UDFSpec]:
-    """List all user-defined function specifications for a user."""
-    statement = select(UDFSpec).where(
-        or_(
-            UDFSpec.owner_id == "tracecat",
-            UDFSpec.owner_id == role.user_id,
-        )
-    )
-    if ns:
-        ns_conds = [UDFSpec.key.startswith(n) for n in ns]
-        statement = statement.where(or_(*ns_conds))
-    if limit:
-        statement = statement.limit(limit)
-    result = session.exec(statement)
-    udfs = result.all()
-    return udfs
-
-
-@app.get("/udfs/{udf_key}", tags=["udfs"])
-def get_udf(
-    role: Annotated[Role, Depends(authenticate_user_or_service)],
-    udf_key: str,
-    namespace: str = Query(None),
-    session: Session = Depends(get_session),
-) -> UDFSpec:
-    """Get a user-defined function specification."""
-    statement = select(UDFSpec).where(
-        or_(
-            UDFSpec.owner_id == "tracecat",
-            UDFSpec.owner_id == role.user_id,
-        ),
-        UDFSpec.key == udf_key,
-    )
-    if namespace:
-        statement = statement.where(UDFSpec.namespace == namespace)
-    result = session.exec(statement)
-    udf = result.one_or_none()
-    if udf is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="udf not found"
-        )
-    return udf
-
-
-@app.post("/udfs/{udf_key}", tags=["udfs"])
-def create_udf(
-    role: Annotated[Role, Depends(authenticate_user)],
-    udf_key: str,
-    session: Session = Depends(get_session),
-) -> UDFSpec:
-    """Create a user-defined function specification."""
-    _, platform, name = udf_key.split(".")
-    statement = select(UDFSpec).where(
-        UDFSpec.owner_id == role.user_id,
-        UDFSpec.platform == platform,
-        UDFSpec.name == name,
-    )
-    result = session.exec(statement)
-    udf = result.one_or_none()
-    if udf is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="udf not found"
-        )
-    return udf
-
-
-@app.post("/udfs/{udf_key}/validate", tags=["udfs"])
-def validate_udf_args(
-    role: Annotated[Role, Depends(authenticate_user)],
-    udf_key: str,
-    args: dict[str, Any],
-) -> UDFArgsValidationResponse:
-    """Validate user-defined function's arguments."""
-    try:
-        result = validation.vadliate_udf_args(udf_key, args)
-        if result.status == "error":
-            logger.error(
-                "Error validating UDF args",
-                message=result.msg,
-                details=result.detail,
-            )
-        return UDFArgsValidationResponse.from_validation_result(result)
-    except KeyError as e:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail=f"UDF {udf_key!r} not found"
-        ) from e
-    except Exception as e:
-        logger.opt(exception=e).error("Error validating UDF args")
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Unexpected error validating UDF args",
-        ) from e
-
-
-@app.post("/validate-workflow", tags=["validation"])
-async def validate_workflow(
-    role: Annotated[Role, Depends(authenticate_user)],
-    definition: UploadFile = File(...),
-    payload: UploadFile = File(None),
-    session: Session = Depends(get_session),
-) -> list[UDFArgsValidationResponse]:
-    """Validate a workflow.
-
-    This deploys the workflow and updates its version. If a YAML file is provided, it will override the workflow in the database."""
-
-    # Committing from YAML (i.e. attaching yaml) will override the workflow definition in the database
-    with logger.contextualize(role=role):
-        # Perform Tiered Validation
-        # Tier 1: DSLInput validation
-        # Verify that the workflow DSL is structurally sound
-        construction_errors = []
-        try:
-            # Uploaded YAML file overrides the workflow in the database
-            dsl = DSLInput.from_yaml(definition.file)
-        except* TracecatValidationError as eg:
-            logger.error(eg.message, error=eg.exceptions)
-            construction_errors.extend(
-                UDFArgsValidationResponse.from_dsl_validation_error(e).model_dump(
-                    exclude_none=True
-                )
-                for e in eg.exceptions
-            )
-        except* ValidationError as eg:
-            logger.error(eg.message, error=eg.exceptions)
-            construction_errors.extend(
-                UDFArgsValidationResponse.from_pydantic_validation_error(e).model_dump(
-                    exclude_none=True
-                )
-                for e in eg.exceptions
-            )
-
-        if construction_errors:
-            msg = f"Workflow definition construction failed with {len(construction_errors)} errors"
-            logger.error(msg)
-            return ORJSONResponse(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                content={
-                    "status": "failure",
-                    "message": msg,
-                    "errors": construction_errors,
-                    "metadata": {"filename": definition.filename}
-                    if definition
-                    else None,
-                },
-            )
-
-        # When we're here, we've verified that the workflow DSL is structurally sound
-        # Now, we have to ensure that the arguments are sound
-
-        if expr_errors := await validation.validate_dsl(session=session, dsl=dsl):
-            msg = f"{len(expr_errors)} validation error(s)"
-            logger.error(msg)
-            return ORJSONResponse(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                content={
-                    "status": "failure",
-                    "message": msg,
-                    "errors": [
-                        UDFArgsValidationResponse.from_validation_result(
-                            val_res
-                        ).model_dump(exclude_none=True)
-                        for val_res in expr_errors
-                    ],
-                    "metadata": {"filename": definition.filename}
-                    if definition
-                    else None,
-                },
-            )
-
-        # Check for input errors
-        if payload:
-            payload_data = orjson.loads(payload.file.read())
-            payload_val_res = validate_trigger_inputs(dsl, payload_data)
-            if payload_val_res.status == "error":
-                msg = "Trigger input validation error"
-                logger.error(msg)
-                return ORJSONResponse(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    content={
-                        "status": "failure",
-                        "message": msg,
-                        "errors": [
-                            UDFArgsValidationResponse.from_validation_result(
-                                payload_val_res
-                            ).model_dump(exclude_none=True)
-                        ],
-                        "metadata": {"filename": definition.filename},
-                    },
-                )
-        return ORJSONResponse(
-            status_code=status.HTTP_200_OK,
-            content={"status": "success", "message": "Workflow passed validation"},
-        )
